[frontend] gui completed

+++ app/controllers/ShoppingListController.scala
@@ -11,7 +11,8 @@ import scala.language.postfixOps
 
 @Singleton
 class ShoppingListController @Inject()(val controllerComponents: ControllerComponents,
-                                       val shoppingListService: ShoppingListService)(implicit ec: ExecutionContext) extends BaseController {
+                                       val shoppingListService: ShoppingListService,
+                                       val authenticatedAction: AuthenticatedAction)(implicit ec: ExecutionContext) extends BaseController {
 
 
   implicit val itemFormatter: OFormat[Item] = Json.format[Item]
@@ -19,69 +20,96 @@ class ShoppingListController @Inject()(val controllerComponents: ControllerCompo
   implicit val shoppingListDtoFormatter: OFormat[ShoppingListDto] = Json.format[ShoppingListDto]
 
 
-  def createShoppingList: Action[AnyContent] = Action.async {
-    shoppingListService.createShoppingList.map {
+  def createShoppingList: Action[AnyContent] = authenticatedAction.async { request =>
+    val userId = request.user.id.get
+    shoppingListService.createShoppingList(userId).map {
       case Some(shoppingList) => Ok(Json.toJson(shoppingList))
       case None => NoContent
     }
   }
 
-  def postShoppingList: Action[ShoppingList] = Action.async(parse.tolerantJson[ShoppingList]) { request =>
+  def postShoppingList: Action[ShoppingList] = authenticatedAction.async(parse.tolerantJson[ShoppingList]) { request =>
+    val userId = request.user.id.get
     val shoppingList = request.body
-    shoppingListService.addShoppingList(shoppingList).map {
+    shoppingListService.addShoppingList(shoppingList, userId).map {
       case Some(savedShoppingList) => Ok(Json.toJson(savedShoppingList))
       case None => NoContent
     }
   }
 
-  def postItemToList(listId: Long): Action[Item] = Action.async(parse.tolerantJson[Item]) { request =>
+  def postItemToList(listId: Long): Action[Item] = authenticatedAction.async(parse.tolerantJson[Item]) { request =>
+    val userId = request.user.id.get
     val item = request.body
-    shoppingListService.addItemToList(listId, item).map {
+    shoppingListService.addItemToList(listId, userId, item).map {
       case Some(savedItem) => Ok(Json.toJson(savedItem))
       case None => NoContent
     }
   }
 
-  def getShoppingList(listId: Long): Action[AnyContent] = Action.async {
-    shoppingListService.getShoppingList(listId).map {
+  def getShoppingList(listId: Long): Action[AnyContent] = authenticatedAction.async { request =>
+    val userId = request.user.id.get
+    shoppingListService.getShoppingList(listId, userId).map {
       case Some(shoppingListDto) => Ok(Json.toJson(shoppingListDto))
       case None => NoContent
     }
   }
 
-  def getShoppingLists: Action[AnyContent] = Action.async {
-    shoppingListService.getShoppingLists.map {
+  def getShoppingLists: Action[AnyContent] = authenticatedAction.async { request =>
+    val userId = request.user.id.get
+    shoppingListService.getShoppingLists(userId).map {
       case Some(shoppingListsDto) => Ok(Json.toJson(shoppingListsDto))
       case None => NoContent
     }
   }
 
-  def getItemWithinList(listId: Long, itemId: Long): Action[AnyContent] = Action.async {
-    shoppingListService.getItemWithinList(listId, itemId).map {
+  def getItemWithinList(listId: Long, itemId: Long): Action[AnyContent] = authenticatedAction.async { request =>
+    val userId = request.user.id.get
+    shoppingListService.getItemWithinList(listId, itemId, userId).map {
       case Some(item) => Ok(Json.toJson(item))
       case None => NoContent
     }
   }
 
-  def getItemsWithinList(listId: Long): Action[AnyContent] = Action.async {
-    shoppingListService.getItemsWithinList(listId).map {
+  def getItemsWithinList(listId: Long): Action[AnyContent] = authenticatedAction.async { request =>
+    val userId = request.user.id.get
+    shoppingListService.getItemsWithinList(listId, userId).map {
       case Some(items) => Ok(Json.toJson(items))
       case None => NoContent
     }
   }
 
-  def deleteShoppingList(listId: Long): Action[AnyContent] = Action.async {
-    shoppingListService.deleteShoppingList(listId).map {
+  def deleteShoppingList(listId: Long): Action[AnyContent] = authenticatedAction.async { request =>
+    val userId = request.user.id.get
+    shoppingListService.deleteShoppingList(listId, userId).map {
       case Some(0) => NotFound
       case Some(_) => Ok
       case None => Conflict
     }
   }
 
-  def deleteItemInList(listId: Long, itemId: Long): Action[AnyContent] = Action.async {
-    shoppingListService.deleteItemInList(listId, itemId).map {
+  def deleteItemInList(listId: Long, itemId: Long): Action[AnyContent] = authenticatedAction.async { request =>
+    val userId = request.user.id.get
+    shoppingListService.deleteItemInList(listId, itemId, userId).map {
       case Some(0) => NotFound
       case Some(_) => Ok
+      case None => NoContent
+    }
+  }
+
+  def putShoppingList(listId: Long): Action[ShoppingList] = authenticatedAction.async(parse.json[ShoppingList]) { request =>
+    val userId = request.user.id.get
+    val shoppingList = request.body
+    shoppingListService.putShoppingList(listId, userId, shoppingList).map {
+      case Some(savedShoppingList) => Ok(Json.toJson(savedShoppingList))
+      case None => NoContent
+    }
+  }
+
+  def putItemToList(listId: Long, itemId: Long): Action[Item] = authenticatedAction.async(parse.json[Item]) { request =>
+    val userId = request.user.id.get
+    val item = request.body
+    shoppingListService.putItemToList(listId, itemId, userId, item).map {
+      case Some(savedItem) => Ok(Json.toJson(savedItem))
       case None => NoContent
     }
   }
